UPDATE: simplifying the file structure and cleaning up code
