+++ src/components/Footer/Footer.jsx
@@ -15,8 +15,8 @@ export const Footer = () => {
   return (
     <SFooter>
       <SFooterTopLines>
-        <img src="./public/assets/svg/lineLeft.svg" alt="lineLeft" />
-        <img src="./public/assets/svg/lineRight.svg" alt="lineRight" />
+        <img src="/assets/svg/lineLeft.svg" alt="lineLeft" />
+        <img src="/assets/svg/lineRight.svg" alt="lineRight" />
       </SFooterTopLines>
       <SFooterInfo>
         <SCompanyLogo>
@@ -26,23 +26,17 @@ export const Footer = () => {
           <SAboutProjectContainer>
             <SAboutProject>პროექტის შესახებ</SAboutProject>
           </SAboutProjectContainer>
-          <img src="./public/assets/png/unilabText.png" alt="unilabText" />
+          <img src="/assets/png/unilabText.png" alt="unilabText" />
         </SAboutProjectDiv>
         <SSocialIcons>
           <a href="">
-            <img
-              src="./public/assets/svg/facebookLogo.svg"
-              alt="facebookLogo"
-            />
+            <img src="/assets/svg/facebookLogo.svg" alt="facebookLogo" />
           </a>
           <a href="">
-            <img src="./public/assets/svg/youtubeLogo.svg" alt="youtubeLogo" />
+            <img src="/assets/svg/youtubeLogo.svg" alt="youtubeLogo" />
           </a>
           <a href="">
-            <img
-              src="./public/assets/svg/linkedinLogo.svg"
-              alt="linkedinLogo"
-            />
+            <img src="/assets/svg/linkedinLogo.svg" alt="linkedinLogo" />
           </a>
         </SSocialIcons>
       </SFooterInfo>
