+++ src/components/Inputs/Checkbox/Check/index.js
@@ -1,0 +1,1 @@
+export { Check } from "./Check";
