+++ src/components/Inputs/Dropdown/Dropdown.jsx
@@ -7,14 +7,9 @@ import {
   SDropdownList,
   SDropdownWrapper,
 } from "./Dropdown.styled";
-import { DropArrow } from "../DropArrow";
+import { DropArrow } from "./DropArrow";
 import { dropdownVariants } from "./Dropdown.variants";
-
-const getLongestString = (strings) => {
-  return strings.reduce((a, b) => {
-    return a.length > b.length ? a : b;
-  });
-};
+import { getLongestString } from "../../../utils/dropdown";
 
 export const Dropdown = ({
   id,
