+++ src/utils/dropdown.js
@@ -1,0 +1,5 @@
+export const getLongestString = (strings) => {
+  return strings.reduce((a, b) => {
+    return a.length > b.length ? a : b;
+  });
+};
